+++ udata/core/dataset/constants.py
@@ -158,7 +158,7 @@ RESOURCE_FILETYPES = OrderedDict(
     ]
 )
 
-OGC_SERVICE_FORMATS = ["wms", "wfs"]
+OGC_SERVICE_FORMATS = ["wfs", "wms", "wmts"]
 
 CHECKSUM_TYPES = ("sha1", "sha2", "sha256", "md5", "crc")
 DEFAULT_CHECKSUM_TYPE = "sha1"