+++ udata/features/notifications/models.py
@@ -3,6 +3,7 @@ from mongoengine import NULLIFY, Q
 from mongoengine.fields import DateTimeField, GenericEmbeddedDocumentField, ReferenceField
 
 from udata.api_fields import field, generate_fields
+from udata.core.dataservices.notifications import DataserviceCreatedNotificationDetails
 from udata.core.discussions.notifications import DiscussionNotificationDetails
 from udata.core.organization.notifications import (
     MembershipAcceptedNotificationDetails,
@@ -10,6 +11,7 @@ from udata.core.organization.notifications import (
     MembershipRequestNotificationDetails,
     NewBadgeNotificationDetails,
 )
+from udata.core.reuse.notifications import ReuseCreatedNotificationDetails
 from udata.core.user.api_fields import user_ref_fields
 from udata.core.user.models import User
 from udata.features.transfer.notifications import TransferRequestNotificationDetails
@@ -72,6 +74,8 @@ class Notification(Datetimed, Document[NotificationQuerySet]):
                 MembershipAcceptedNotificationDetails,
                 MembershipRefusedNotificationDetails,
                 ValidateHarvesterNotificationDetails,
+                ReuseCreatedNotificationDetails,
+                DataserviceCreatedNotificationDetails,
             )
         ),
         generic=True,